# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": [
          "training-session-controller"
        ],
        "operationId": "getAllSessions",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "training-session-controller"
        ],
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSession"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/participants/import/session/{sessionId}": {
      "post": {
        "tags": [
          "participant-controller"
        ],
        "operationId": "importParticipantsFromSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/{id}/retry": {
      "post": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "retryNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/send": {
      "post": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "sendNotification",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sessions/{id}/status": {
      "patch": {
        "tags": [
          "training-session-controller"
        ],
        "operationId": "updateSessionStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "training-session-controller"
        ],
        "operationId": "getSessionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingSessionDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "subsidiaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subsidiary"
            },
            "uniqueItems": true
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "action": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleName": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "OPERATIONAL_ADMIN",
              "TRAINING_MANAGER",
              "TRAINER",
              "READ_ONLY"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "uniqueItems": true
          }
        }
      },
      "Subsidiary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          }
        }
      },
      "Trainer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "specialty": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "affiliatedOrganizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Organization"
            },
            "uniqueItems": true
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PENDING"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "Attendance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "attendanceStatus": {
            "type": "string",
            "enum": [
              "PRESENT",
              "LATE",
              "ABSENT",
              "JUSTIFIED_ABSENCE"
            ]
          },
          "sessionParticipant": {
            "$ref": "#/components/schemas/SessionParticipant"
          }
        }
      },
      "Participant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "registrationSource": {
            "type": "string"
          }
        }
      },
      "SessionParticipant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "trainingSession": {
            "$ref": "#/components/schemas/TrainingSession"
          },
          "participant": {
            "$ref": "#/components/schemas/Participant"
          },
          "attendances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attendance"
            }
          }
        }
      },
      "TrainingSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "onSite": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "CANCELLED",
              "SCHEDULED",
              "ONGOING",
              "POSTPONED"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "ON_SITE",
              "REMOTE",
              "HYBRID"
            ]
          },
          "trainer": {
            "$ref": "#/components/schemas/Trainer"
          },
          "subsidiary": {
            "$ref": "#/components/schemas/Subsidiary"
          },
          "sessionParticipants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionParticipant"
            }
          }
        }
      },
...
      "TrainingSessionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "onSite": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "CANCELLED",
              "SCHEDULED",
              "ONGOING",
              "POSTPONED"
            ]
          },
          "mode": {
            "type": "string",
            "enum": [
              "ON_SITE",
              "REMOTE",
              "HYBRID"
            ]
          },
          "trainer": {
            "$ref": "#/components/schemas/Trainer"
          },
          "subsidiary": {
            "$ref": "#/components/schemas/Subsidiary"
          },
          "sessionParticipants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionParticipant"
            }
          }
        }
      },
...
    }
  }
}